Size the public counters in Go, not in the dependencies' types
ConnPoolMetrics carried puddle's int32 gauges and hand-widened uint64
counters, with the acquire wait time as raw nanoseconds; BreakerConfig
and BreakerStats carried gobreaker's uint32. Use database/sql.DBStats'
shape instead: int gauges, int64 lifetime counters, a time.Duration for
the wait, and int for the breaker's counts and thresholds.

A negative TripMinRequests or HalfOpenMaxRequests now takes the default,
like the ratio and the durations.

+++ loadtest/cmd/loadgen/main.go
@@ -90,7 +90,7 @@ func main() {
 
 	clientConfig := prof.ClientConfig()
 	if *breakerTrip > 0 {
-		trip := uint32(*breakerTrip)
+		trip := *breakerTrip
 		clientConfig.Breaker = breakerConfig(trip, *breakerOpen)
 		log.Info("circuit breaker enabled", "trip_after", trip, "open_interval", *breakerOpen)
 	}
@@ -198,7 +198,7 @@ loop:
 	}
 }
 
-func breakerConfig(minRequests uint32, openDuration time.Duration) memcache.BreakerConfig {
+func breakerConfig(minRequests int, openDuration time.Duration) memcache.BreakerConfig {
 	return memcache.BreakerConfig{
 		Enabled:          true,
 		TripMinRequests:  minRequests,
